Big commit after project split

+++ dynamo-core/src/main/java/dynamo/backlog/tasks/core/CancelDownloadTaskExecutor.java
@@ -1,9 +1,9 @@
 package dynamo.backlog.tasks.core;
 
+import dynamo.backlog.tasks.torrent.Transmission;
 import dynamo.core.model.TaskExecutor;
 import dynamo.jdbi.SearchResultDAO;
 import dynamo.model.result.SearchResultType;
-import dynamo.torrents.transmission.Transmission;
 import dynamo.webapps.sabnzbd.SabNzbd;
 
 public class CancelDownloadTaskExecutor extends TaskExecutor<CancelDownloadTask> {
@@ -19,7 +19,7 @@ public class CancelDownloadTaskExecutor extends TaskExecutor<CancelDownloadTask>
 	public void execute() throws Exception {
 		// hackish : shouldn't be implemented like this
 		if ( task.getResult().getType() == SearchResultType.TORRENT && Transmission.getInstance().isEnabled()) {
-			Transmission.getInstance().remove( Integer.parseInt( task.getResult().getClientId() ), true );
+			Transmission.getInstance().remove( Long.parseLong( task.getResult().getClientId() ), true );
 		}
 		else if ( task.getResult().getType() == SearchResultType.NZB && SabNzbd.getInstance().isEnabled()) {
 			SabNzbd.getInstance().remove( task.getResult().getClientId() );
